requirements.txt
sentence-transformers==2.2.2
pypdf==3.8.1
pygpt4all==1.1.0
pdf2image==1.16.3
faiss-cpu
langchain==0.0.228
gpt4all==1.0.3
python-dotenv==1.0.0
